% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apprunner_operations.R
\name{apprunner_delete_auto_scaling_configuration}
\alias{apprunner_delete_auto_scaling_configuration}
\title{Delete an App Runner automatic scaling configuration resource}
\usage{
apprunner_delete_auto_scaling_configuration(AutoScalingConfigurationArn)
}
\arguments{
\item{AutoScalingConfigurationArn}{[required] The Amazon Resource Name (ARN) of the App Runner auto scaling
configuration that you want to delete.

The ARN can be a full auto scaling configuration ARN, or a partial ARN
ending with either \code{.../name } or \code{.../name/revision }. If a revision
isn't specified, the latest active revision is deleted.}
}
\description{
Delete an App Runner automatic scaling configuration resource. You can delete a specific revision or the latest active revision. You can't delete a configuration that's used by one or more App Runner services.

See \url{https://paws-r.github.io/docs/apprunner/delete_auto_scaling_configuration.html} for full documentation.
}
\keyword{internal}
