% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braket_operations.R
\name{braket_search_jobs}
\alias{braket_search_jobs}
\title{Searches for Amazon Braket jobs that match the specified filter values}
\usage{
braket_search_jobs(filters, maxResults = NULL, nextToken = NULL)
}
\arguments{
\item{filters}{[required] The filter values to use when searching for a job.}

\item{maxResults}{The maximum number of results to return in the response.}

\item{nextToken}{A token used for pagination of results returned in the response. Use the
token returned from the previous request to continue results where the
previous request ended.}
}
\description{
Searches for Amazon Braket jobs that match the specified filter values.

See \url{https://paws-r.github.io/docs/braket/search_jobs.html} for full documentation.
}
\keyword{internal}
