% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecr_operations.R
\name{ecr_get_registry_policy}
\alias{ecr_get_registry_policy}
\title{Retrieves the permissions policy for a registry}
\usage{
ecr_get_registry_policy()
}
\description{
Retrieves the permissions policy for a registry.

See \url{https://paws-r.github.io/docs/ecr/get_registry_policy.html} for full documentation.
}
\keyword{internal}
