% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecrpublic_operations.R
\name{ecrpublic_initiate_layer_upload}
\alias{ecrpublic_initiate_layer_upload}
\title{Notifies Amazon ECR that you intend to upload an image layer}
\usage{
ecrpublic_initiate_layer_upload(registryId = NULL, repositoryName)
}
\arguments{
\item{registryId}{The AWS account ID associated with the registry to which you intend to
upload layers. If you do not specify a registry, the default public
registry is assumed.}

\item{repositoryName}{[required] The name of the repository to which you intend to upload layers.}
}
\description{
Notifies Amazon ECR that you intend to upload an image layer.

See \url{https://paws-r.github.io/docs/ecrpublic/initiate_layer_upload.html} for full documentation.
}
\keyword{internal}
