% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecrpublic_operations.R
\name{ecrpublic_tag_resource}
\alias{ecrpublic_tag_resource}
\title{Associates the specified tags to a resource with the specified
resourceArn}
\usage{
ecrpublic_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource to which to add tags.
Currently, the supported resource is an Amazon ECR Public repository.}

\item{tags}{[required] The tags to add to the resource. A tag is an array of key-value pairs.
Tag keys can have a maximum character length of 128 characters, and tag
values can have a maximum length of 256 characters.}
}
\description{
Associates the specified tags to a resource with the specified \code{resourceArn}. If existing tags on a resource are not specified in the request parameters, they are not changed. When a resource is deleted, the tags associated with that resource are deleted as well.

See \url{https://paws-r.github.io/docs/ecrpublic/tag_resource.html} for full documentation.
}
\keyword{internal}
