% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecs_operations.R
\name{ecs_update_cluster_settings}
\alias{ecs_update_cluster_settings}
\title{Modifies the settings to use for a cluster}
\usage{
ecs_update_cluster_settings(cluster, settings)
}
\arguments{
\item{cluster}{[required] The name of the cluster to modify the settings for.}

\item{settings}{[required] The setting to use by default for a cluster. This parameter is used to
turn on CloudWatch Container Insights for a cluster. If this value is
specified, it overrides the \code{containerInsights} value set with
\code{\link[=ecs_put_account_setting]{put_account_setting}} or
\code{\link[=ecs_put_account_setting_default]{put_account_setting_default}}.}
}
\description{
Modifies the settings to use for a cluster.

See \url{https://paws-r.github.io/docs/ecs/update_cluster_settings.html} for full documentation.
}
\keyword{internal}
