% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eks_operations.R
\name{eks_associate_encryption_config}
\alias{eks_associate_encryption_config}
\title{Associate encryption configuration to an existing cluster}
\usage{
eks_associate_encryption_config(
  clusterName,
  encryptionConfig,
  clientRequestToken = NULL
)
}
\arguments{
\item{clusterName}{[required] The name of the cluster that you are associating with encryption
configuration.}

\item{encryptionConfig}{[required] The configuration you are using for encryption.}

\item{clientRequestToken}{The client request token you are using with the encryption
configuration.}
}
\description{
Associate encryption configuration to an existing cluster.

See \url{https://paws-r.github.io/docs/eks/associate_encryption_config.html} for full documentation.
}
\keyword{internal}
