% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eks_operations.R
\name{eks_delete_addon}
\alias{eks_delete_addon}
\title{Delete an Amazon EKS add-on}
\usage{
eks_delete_addon(clusterName, addonName, preserve = NULL)
}
\arguments{
\item{clusterName}{[required] The name of the cluster to delete the add-on from.}

\item{addonName}{[required] The name of the add-on. The name must match one of the names returned by
\href{https://docs.aws.amazon.com/eks/latest/APIReference/API_ListAddons.html}{\code{list_addons}}
.}

\item{preserve}{Specifying this option preserves the add-on software on your cluster but
Amazon EKS stops managing any settings for the add-on. If an IAM account
is associated with the add-on, it is not removed.}
}
\description{
Delete an Amazon EKS add-on.

See \url{https://paws-r.github.io/docs/eks/delete_addon.html} for full documentation.
}
\keyword{internal}
