% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eks_operations.R
\name{eks_delete_cluster}
\alias{eks_delete_cluster}
\title{Deletes the Amazon EKS cluster control plane}
\usage{
eks_delete_cluster(name)
}
\arguments{
\item{name}{[required] The name of the cluster to delete.}
}
\description{
Deletes the Amazon EKS cluster control plane.

See \url{https://paws-r.github.io/docs/eks/delete_cluster.html} for full documentation.
}
\keyword{internal}
