% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eks_operations.R
\name{eks_register_cluster}
\alias{eks_register_cluster}
\title{Connects a Kubernetes cluster to the Amazon EKS control plane}
\usage{
eks_register_cluster(
  name,
  connectorConfig,
  clientRequestToken = NULL,
  tags = NULL
)
}
\arguments{
\item{name}{[required] Define a unique name for this cluster for your Region.}

\item{connectorConfig}{[required] The configuration settings required to connect the Kubernetes cluster to
the Amazon EKS control plane.}

\item{clientRequestToken}{Unique, case-sensitive identifier that you provide to ensure the
idempotency of the request.}

\item{tags}{The metadata that you apply to the cluster to assist with categorization
and organization. Each tag consists of a key and an optional value, both
of which you define. Cluster tags do not propagate to any other
resources associated with the cluster.}
}
\description{
Connects a Kubernetes cluster to the Amazon EKS control plane.

See \url{https://paws-r.github.io/docs/eks/register_cluster.html} for full documentation.
}
\keyword{internal}
