% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emrcontainers_operations.R
\name{emrcontainers_describe_managed_endpoint}
\alias{emrcontainers_describe_managed_endpoint}
\title{Displays detailed information about a managed endpoint}
\usage{
emrcontainers_describe_managed_endpoint(id, virtualClusterId)
}
\arguments{
\item{id}{[required] This output displays ID of the managed endpoint.}

\item{virtualClusterId}{[required] The ID of the endpoint's virtual cluster.}
}
\description{
Displays detailed information about a managed endpoint. A managed endpoint is a gateway that connects EMR Studio to Amazon EMR on EKS so that EMR Studio can communicate with your virtual cluster.

See \url{https://paws-r.github.io/docs/emrcontainers/describe_managed_endpoint.html} for full documentation.
}
\keyword{internal}
