% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_create_distribution_configuration}
\alias{imagebuilder_create_distribution_configuration}
\title{Creates a new distribution configuration}
\usage{
imagebuilder_create_distribution_configuration(
  name,
  description = NULL,
  distributions,
  tags = NULL,
  clientToken
)
}
\arguments{
\item{name}{[required] The name of the distribution configuration.}

\item{description}{The description of the distribution configuration.}

\item{distributions}{[required] The distributions of the distribution configuration.}

\item{tags}{The tags of the distribution configuration.}

\item{clientToken}{[required] The idempotency token of the distribution configuration.}
}
\description{
Creates a new distribution configuration. Distribution configurations define and configure the outputs of your pipeline.

See \url{https://paws-r.github.io/docs/imagebuilder/create_distribution_configuration.html} for full documentation.
}
\keyword{internal}
