% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_delete_image_recipe}
\alias{imagebuilder_delete_image_recipe}
\title{Deletes an image recipe}
\usage{
imagebuilder_delete_image_recipe(imageRecipeArn)
}
\arguments{
\item{imageRecipeArn}{[required] The Amazon Resource Name (ARN) of the image recipe to delete.}
}
\description{
Deletes an image recipe.

See \url{https://paws-r.github.io/docs/imagebuilder/delete_image_recipe.html} for full documentation.
}
\keyword{internal}
