% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_update_distribution_configuration}
\alias{imagebuilder_update_distribution_configuration}
\title{Updates a new distribution configuration}
\usage{
imagebuilder_update_distribution_configuration(
  distributionConfigurationArn,
  description = NULL,
  distributions,
  clientToken
)
}
\arguments{
\item{distributionConfigurationArn}{[required] The Amazon Resource Name (ARN) of the distribution configuration that
you want to update.}

\item{description}{The description of the distribution configuration.}

\item{distributions}{[required] The distributions of the distribution configuration.}

\item{clientToken}{[required] The idempotency token of the distribution configuration.}
}
\description{
Updates a new distribution configuration. Distribution configurations define and configure the outputs of your pipeline.

See \url{https://paws-r.github.io/docs/imagebuilder/update_distribution_configuration.html} for full documentation.
}
\keyword{internal}
