% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proton_operations.R
\name{proton_get_service}
\alias{proton_get_service}
\title{Get detailed data for a service}
\usage{
proton_get_service(name)
}
\arguments{
\item{name}{[required] The name of the service that you want to get the detailed data for.}
}
\description{
Get detailed data for a service.

See \url{https://paws-r.github.io/docs/proton/get_service.html} for full documentation.
}
\keyword{internal}
