% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proton_operations.R
\name{proton_list_service_instances}
\alias{proton_list_service_instances}
\title{List service instances with summary data}
\usage{
proton_list_service_instances(
  maxResults = NULL,
  nextToken = NULL,
  serviceName = NULL
)
}
\arguments{
\item{maxResults}{The maximum number of service instances to list.}

\item{nextToken}{A token that indicates the location of the next service in the array of
service instances, after the list of service instances that was
previously requested.}

\item{serviceName}{The name of the service that the service instance belongs to.}
}
\description{
List service instances with summary data.

See \url{https://paws-r.github.io/docs/proton/list_service_instances.html} for full documentation.
}
\keyword{internal}
