% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proton_operations.R
\name{proton_reject_environment_account_connection}
\alias{proton_reject_environment_account_connection}
\title{In a management account, reject an environment account connection from
another environment account}
\usage{
proton_reject_environment_account_connection(id)
}
\arguments{
\item{id}{[required] The ID of the environment account connection to reject.}
}
\description{
In a management account, reject an environment account connection from another environment account.

See \url{https://paws-r.github.io/docs/proton/reject_environment_account_connection.html} for full documentation.
}
\keyword{internal}
