% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proton_operations.R
\name{proton_update_account_settings}
\alias{proton_update_account_settings}
\title{Update the Proton service pipeline role or repository settings}
\usage{
proton_update_account_settings(
  pipelineProvisioningRepository = NULL,
  pipelineServiceRoleArn = NULL
)
}
\arguments{
\item{pipelineProvisioningRepository}{A repository for pipeline provisioning. Specify it if you have
environments configured for self-managed provisioning with services that
include pipelines.}

\item{pipelineServiceRoleArn}{The Amazon Resource Name (ARN) of the service role you want to use for
provisioning pipelines. Assumed by Proton for Amazon Web
Services-managed provisioning, and by customer-owned automation for
self-managed provisioning.}
}
\description{
Update the Proton service pipeline role or repository settings.

See \url{https://paws-r.github.io/docs/proton/update_account_settings.html} for full documentation.
}
\keyword{internal}
